+++ vscode-extension/src/commands/collapse-all-rangeExplorer.ts
@@ -1,0 +1,6 @@
+import { TreeViews } from "../features/Explorer/TreeViews";
+
+export function collapseAllRangeExplorer() {
+    const controller = TreeViews.instance.getExpandCollapse("ninja-rangeExplorer");
+    controller?.collapseAll();
+}