+++ vscode-extension/src/features/Explorer/TextTreeItem.ts
@@ -4,11 +4,13 @@ import { INinjaTreeItem } from "./NinjaTreeItem";
 export class TextTreeItem implements INinjaTreeItem {
     private readonly _text: string;
     private readonly _tooltip: string;
+    private readonly _id: string;
 
     constructor(text: string, tooltip: string, parent: INinjaTreeItem | undefined) {
         this.parent = parent;
         this._text = text;
         this._tooltip = tooltip;
+        this._id = `${parent?.id || ""}.[${text}]`;
     }
 
     public readonly parent: INinjaTreeItem | undefined;
@@ -19,5 +21,9 @@ export class TextTreeItem implements INinjaTreeItem {
         const item = new TreeItem(this._text, TreeItemCollapsibleState.None);
         item.tooltip = this._tooltip;
         return item;
+    }
+
+    public get id() {
+        return this._id;
     }
 }